Refining key points, and a bit more
Some suggestions for the plotting episode. I squeezed in accessibility considerations, and tweaked the exercise that critiqued the figure. Some of the previous critiques that would likely require exporting the plot to Adobe Illustrator, which is beyond the scope of the lesson and probably not a useful thing to do in most cases.